Fix WP Hook Errors Fixes #26
